Tidying up
Adding xml comments, moving temp stuff into the temp folder, rename EventProcessorIdentifier to EventProcessorId
