+++ Source/Events/Processing/ScopedEventProcessor.cs
@@ -4,6 +4,7 @@ namespace Dolittle.Runtime.Events.Processing
     using System.Collections.Concurrent;
     using System.Collections.Generic;
     using System.Threading.Tasks;
+    using Dolittle.Artifacts;
     using Dolittle.Bootstrapping;
     using Dolittle.Logging;
     using Dolittle.Runtime.Events;
@@ -12,19 +13,32 @@ namespace Dolittle.Runtime.Events.Processing
     using Dolittle.Types;
 
     /// <summary>
-    /// 
+    /// Processes an individiual <see cref="CommittedEventEnvelope" /> for the correct <see cref="TenantId" />
     /// </summary>
     public class ScopedEventProcessor
     {
         CommittedEventVersion _lastVersionProcessed;
 
+        /// <summary>
+        /// A <see cref="ScopedEventProcessorKey" /> to identity the <see cref="Artifact">Event</see> and <see cref="TenantId">Tenant</see> combination
+        /// </summary>
+        /// <value></value>
         public ScopedEventProcessorKey Key { get; }
-        public EventProcessorIdentifier ProcessorId => _processor?.Identifier;
+        /// <summary>
+        /// Unique identifer for the <see cref="IEventProcessor" />
+        /// </summary>
+        public EventProcessorId ProcessorId => _processor?.Identifier;
 
         TenantId _tenant;
         IEventProcessor _processor;
         ILogger _logger;
 
+        /// <summary>
+        /// Instantiates an instance of <see cref="ScopedEventProcessor" />
+        /// </summary>
+        /// <param name="tenant">The <see cref="TenantId" /> that this processor is scoped to.</param>
+        /// <param name="processor">An <see cref="IEventProcessor" /> to process the event</param>
+        /// <param name="logger">An <see cref="ILogger" /> to log messages</param>
         public ScopedEventProcessor(TenantId tenant, IEventProcessor processor, ILogger logger)
         {
             _tenant = tenant;
@@ -33,6 +47,9 @@ namespace Dolittle.Runtime.Events.Processing
             Key = new ScopedEventProcessorKey(tenant,processor.Event);
         }
 
+        /// <summary>
+        /// Instructs the Processor to Catch up by processing events that have been committed since the last processed event
+        /// </summary>
         public virtual void CatchUp()
         {
             //SetIsCaughtUp(committedEventVersion);
@@ -44,6 +61,10 @@ namespace Dolittle.Runtime.Events.Processing
             HasCaughtUp = true;
         }
 
+        /// <summary>
+        /// Processes a single <see cref="CommittedEventEnvelope" />
+        /// </summary>
+        /// <param name="envelope">The <see cref="CommittedEventEnvelope" />to process</param>
         public virtual void Process(CommittedEventEnvelope envelope)
         {
             if(CanProcess(envelope.Version))
@@ -55,8 +76,18 @@ namespace Dolittle.Runtime.Events.Processing
             _lastVersionProcessed = envelope.Version;
             _logger.Debug($"Processing {envelope.Version}");
         }
+
+        /// <summary>
+        /// Indicates whether the processor has caught up with committed events
+        /// </summary>
+        /// <value></value>
         public bool HasCaughtUp { get; private set; }
 
+        /// <summary>
+        /// Indicates whether the processor can process the <see cref="CommittedEventVersion" /> provided.
+        /// </summary>
+        /// <param name="version"></param>
+        /// <returns>True if the processor has caught up and the version is greater than the last processed version, otherwise false</returns>
         public bool CanProcess(CommittedEventVersion version)
         {
             return HasCaughtUp && version > _lastVersionProcessed;
