+++ Source/Events/Processing/Temp/IScopedEventOffsetProvider.cs
@@ -1,21 +1,27 @@
+#pragma warning disable 1591
+
 namespace Dolittle.Runtime.Events.Processing
 {
+    using Dolittle.Runtime.Tenancy;
+
     /// <summary>
     /// 
     /// </summary>
-    public interface IEventOffsetProvider
+    public interface IScopedEventOffsetProvider 
     {
         /// <summary>
         /// 
         /// </summary>
         /// <param name="id"></param>
+        /// <param name="tenant"></param>
         /// <param name="offset"></param>
-        void Set(EventProcessorIdentifier id, ulong offset);
+        void Set(EventProcessorId id, TenantId tenant, ulong offset);
         /// <summary>
         /// 
         /// </summary>
         /// <param name="id"></param>
+        /// <param name="tenant"></param>
         /// <returns></returns>
-        ulong Get(EventProcessorIdentifier id);
+        ulong Get(EventProcessorId id, TenantId tenant);
     }
 }
