+++ Source/Events/Processing/Temp/ITenants.cs
@@ -1,3 +1,5 @@
+#pragma warning disable 1591
+
 namespace Dolittle.Runtime.Events.Processing
 {
     using System.Collections.Generic;
