+++ Specifications/Events/for_EventSourceVersion/when_incrementing_the_sequence/on_no_version.cs
@@ -6,7 +6,6 @@ namespace Dolittle.Runtime.Events.for_EventSourceVersion.when_incrementing_the_s
     [Subject(typeof(EventSourceVersion),"NextSequence")]
     public class on_no_version
     {
-        static EventSourceVersion result;
         static Exception exception;
 
         Because of = () => exception = Catch.Exception(() => EventSourceVersion.NoVersion.NextSequence());
